requirements.txt
nsetools
python-telegram-bot
